// react-cpa/src/pages/Competitor/HelpfulReview/helpfulReview.d.ts
/*
 * @Author: wjw
 * @Date: 2020-11-16 16:34:19
 * @LastEditTime: 2020-11-17 14:36:32
 * @LastEditors: Please set LastEditors
 * @Description: In User Settings Edit
 * @FilePath: \react-punkd:\work\react-competitor\src\pages\Competitor\HelpfulReview\helpfulReview.d.ts
 */

export interface HelpfulReviewData {
  proportion: {
    five: number;
    four: number;
    three: number;
    two: number;
    one: number;
    all: number;
  };
  weight: {
    five: number;
    four: number;
    three: number;
    two: number;
    one: number;
    all: number;
  };
}